requirements.txt
flask
gunicorn
requests
requests-oauthlib
psycopg2-binary
certifi
python-dotenv==1.0.0
